chg: usr: raises if activation vector file exists

+++ ruleskit/activation.py
@@ -149,6 +149,8 @@ class Activation(ABC):
         t0 = time()
         self.data = Activation.DEFAULT_TEMPDIR / f"ACTIVATION_VECTOR_{name}.txt"
         self.data_format = "file"
+        if self.data.isfile():
+            raise FileExistsError(f"There is already an activation vector with the name {name}")
         with open(self.data, "wb") as f:
             # noinspection PyTypeChecker
             np.save(f, value, allow_pickle=False)
